Moved views into util.py
