+++ corehq/apps/export/old_views.py
@@ -23,6 +23,7 @@ from corehq.util.timezones.utils import get_timezone_for_user
 from corehq.toggles import MESSAGE_LOG_METADATA, PAGINATED_EXPORTS
 from corehq.apps.export.export import get_export_download, get_export_size
 from corehq.apps.export.models.new import DatePeriod, DataFile, EmailExportWhenDoneRequest
+from corehq.apps.export.views.util import DailySavedExportMixin, DailySavedExportMixin
 from corehq.apps.hqwebapp.views import HQJSONResponseMixin
 from corehq.apps.hqwebapp.utils import format_angular_error, format_angular_success
 from corehq.apps.locations.models import SQLLocation
@@ -142,64 +143,6 @@ def user_can_view_deid_exports(domain, couch_user):
                 get_permission_name(Permissions.view_report),
                 data=DEID_EXPORT_PERMISSION
             ))
-
-
-class ExportsPermissionsManager(object):
-    """
-    Encapsulates some shortcuts for checking export permissions.
-
-    Users need to have edit permissions to create or update exports
-    Users need the "view reports" permission to download exports
-    The DEIDENTIFIED_DATA privilege is a pro-plan feature, and without it,
-        users should not be able to create, update, or download deid exports.
-    There are some users with access to a specific DeidExportReport.  If these
-        users do not have the "view reports" permission, they should only be
-        able to access deid reports.
-    """
-
-    def __init__(self, form_or_case, domain, couch_user):
-        super(ExportsPermissionsManager, self).__init__()
-        if form_or_case not in [None, 'form', 'case']:
-            raise ValueError("Unrecognized value for form_or_case")
-        self.form_or_case = form_or_case
-        self.domain = domain
-        self.couch_user = couch_user
-
-    @property
-    def has_edit_permissions(self):
-        return self.couch_user.can_edit_data()
-
-    @property
-    def has_form_export_permissions(self):
-        return has_permission_to_view_report(self.couch_user, self.domain, FORM_EXPORT_PERMISSION)
-
-    @property
-    def has_case_export_permissions(self):
-        return has_permission_to_view_report(self.couch_user, self.domain, CASE_EXPORT_PERMISSION)
-
-    @property
-    def has_view_permissions(self):
-        if self.form_or_case is None:
-            return self.has_form_export_permissions or self.has_case_export_permissions
-        elif self.form_or_case == "form":
-            return self.has_form_export_permissions
-        elif self.form_or_case == "case":
-            return self.has_case_export_permissions
-        return False
-
-    @property
-    def has_deid_view_permissions(self):
-        # just a convenience wrapper around user_can_view_deid_exports
-        return user_can_view_deid_exports(self.domain, self.couch_user)
-
-    def access_list_exports_or_404(self, is_deid=False):
-        if not (self.has_edit_permissions or self.has_view_permissions
-                or (is_deid and self.has_deid_view_permissions)):
-            raise Http404()
-
-    def access_download_export_or_404(self):
-        if not (self.has_edit_permissions or self.has_view_permissions or self.has_deid_view_permissions):
-            raise Http404()
 
 
 @location_safe
@@ -489,65 +432,6 @@ class CreateNewCustomCaseExportView(BaseModifyNewCustomView):
         return super(CreateNewCustomCaseExportView, self).get(request, *args, **kwargs)
 
 
-class DailySavedExportMixin(object):
-
-    def _priv_check(self):
-        if not domain_has_privilege(self.domain, DAILY_SAVED_EXPORT):
-            raise Http404
-
-    def dispatch(self, *args, **kwargs):
-        self._priv_check()
-        return super(DailySavedExportMixin, self).dispatch(*args, **kwargs)
-
-    def create_new_export_instance(self, schema):
-        instance = super(DailySavedExportMixin, self).create_new_export_instance(schema)
-        instance.is_daily_saved_export = True
-
-        span = datespan_from_beginning(self.domain_object, _get_timezone(self.domain, self.request.couch_user))
-        instance.filters.date_period = DatePeriod(
-            period_type="since", begin=span.startdate.date()
-        )
-        if not self.request.can_access_all_locations:
-            accessible_location_ids = (SQLLocation.active_objects.accessible_location_ids(
-                self.request.domain,
-                self.request.couch_user)
-            )
-        else:
-            accessible_location_ids = None
-        instance.filters.can_access_all_locations = self.request.can_access_all_locations
-        instance.filters.accessible_location_ids = accessible_location_ids
-
-        return instance
-
-    @property
-    def report_class(self):
-        from corehq.apps.export.views.list import DailySavedExportListView
-        return DailySavedExportListView
-
-
-class DashboardFeedMixin(DailySavedExportMixin):
-
-    def _priv_check(self):
-        if not domain_has_privilege(self.domain, EXCEL_DASHBOARD):
-            raise Http404
-
-    def create_new_export_instance(self, schema):
-        instance = super(DashboardFeedMixin, self).create_new_export_instance(schema)
-        instance.export_format = "html"
-        return instance
-
-    @property
-    def page_context(self):
-        context = super(DashboardFeedMixin, self).page_context
-        context['format_options'] = ["html"]
-        return context
-
-    @property
-    def report_class(self):
-        from corehq.apps.export.views.list import DashboardFeedListView
-        return DashboardFeedListView
-
-
 @location_safe
 class CreateNewCaseFeedView(DashboardFeedMixin, CreateNewCustomCaseExportView):
     urlname = 'new_case_feed_export'
@@ -701,47 +585,6 @@ class DeleteNewCustomExportView(BaseModifyNewCustomView):
             return CaseExportListView
         else:
             raise Exception("Export does not match any export list views!")
-
-
-class GenerateSchemaFromAllBuildsView(View):
-    urlname = 'build_full_schema'
-
-    def export_cls(self, type_):
-        return CaseExportDataSchema if type_ == CASE_EXPORT else FormExportDataSchema
-
-    def get(self, request, *args, **kwargs):
-        download_id = request.GET.get('download_id')
-        download = DownloadBase.get(download_id)
-        if download is None:
-            return json_response({
-                'download_id': download_id,
-                'progress': None,
-            })
-
-        status = get_task_status(download.task)
-        return json_response({
-            'download_id': download_id,
-            'success': status.success(),
-            'failed': status.failed(),
-            'missing': status.missing(),
-            'not_started': status.not_started(),
-            'progress': status.progress._asdict(),
-        })
-
-    def post(self, request, *args, **kwargs):
-        type_ = request.POST.get('type')
-        assert type_ in [CASE_EXPORT, FORM_EXPORT], 'Unrecogized export type {}'.format(type_)
-        download = DownloadBase()
-        download.set_task(generate_schema_for_all_builds.delay(
-            self.export_cls(type_),
-            request.domain,
-            request.POST.get('app_id'),
-            request.POST.get('identifier'),
-        ))
-        download.save()
-        return json_response({
-            'download_id': download.download_id
-        })
 
 
 def can_download_daily_saved_export(export, domain, couch_user):
