+++ corehq/apps/export/views/list.py
@@ -23,7 +23,7 @@ from corehq.util.timezones.utils import get_timezone_for_user
 from corehq.toggles import MESSAGE_LOG_METADATA, PAGINATED_EXPORTS
 from corehq.apps.export.export import get_export_download, get_export_size
 from corehq.apps.export.models.new import DatePeriod, DataFile, EmailExportWhenDoneRequest
-from corehq.apps.export.old_views import ExportsPermissionsManager
+from corehq.apps.export.views.util import ExportsPermissionsManager
 from corehq.apps.hqwebapp.views import HQJSONResponseMixin
 from corehq.apps.hqwebapp.utils import format_angular_error, format_angular_success
 from corehq.apps.locations.models import SQLLocation
